# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /users:
    # binds a127 app logic to a route
    x-swagger-router-controller: users
    get:
      description: Returns all users to the caller
      x-swagger-serializer: users
      # used as the method name of the controller and serializer
      operationId: usersGet
      parameters:
        - name: name
          in: query
          description: User name to get
          required: false
          type: string
        - name: email
          in: query
          description: User email to get
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            type: array
            items:
              $ref: "#/definitions/User"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Create a user
      x-swagger-deserializer: users
      operationId: usersPost
      parameters:
        - name: user
          in: body
          description: Properties of new user
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "201":
          description: User created   
          schema:
            type: object
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse" 
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  User:
    type: object
    properties:
      name:
        type: string
      email:
        type: string
    required:
      - name
      - email
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string